refactor: extract story detail logic and components
